+++ fabric/src/main/java/net/satisfy/candlelight/fabric/client/renderer/DyeableCandlelightLeggingsRenderer.java
@@ -15,21 +15,22 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.component.CustomData;
 import net.satisfy.candlelight.core.item.DyeableCandlelightArmorItem;
 import net.satisfy.candlelight.core.registry.ArmorRegistry;
-import net.satisfy.candlelight.core.util.CandlelightIdentifier;
 
 public class DyeableCandlelightLeggingsRenderer implements ArmorRenderer {
     @Override
     public void render(PoseStack matrices, MultiBufferSource vertexConsumers, ItemStack stack, LivingEntity entity, EquipmentSlot slot, int light, HumanoidModel<LivingEntity> contextModel) {
-        if (!(stack.getItem() instanceof DyeableCandlelightArmorItem armorItem)) return;
+        if (slot != EquipmentSlot.LEGS) return;
+        if (!(stack.getItem() instanceof DyeableCandlelightArmorItem item)) return;
         CompoundTag tag = stack.getOrDefault(DataComponents.CUSTOM_DATA, CustomData.EMPTY).copyTag();
-
         if (tag.contains("Visible") && !tag.getBoolean("Visible")) return;
-        Model model = ArmorRegistry.getSuitModel(armorItem, contextModel.rightLeg, contextModel.leftLeg);
-
-        int packedColor = 0xFF000000 | armorItem.getColor(stack);
-
-        model.renderToBuffer(matrices, vertexConsumers.getBuffer(model.renderType(armorItem.getTexture())), light, OverlayTexture.NO_OVERLAY, packedColor);
-        ResourceLocation overlayTexture = CandlelightIdentifier.identifier("textures/models/armor/dress_overlay.png");
-        model.renderToBuffer(matrices, vertexConsumers.getBuffer(model.renderType(overlayTexture)), light, OverlayTexture.NO_OVERLAY, 0xFFFFFFFF);
+        Model model = ArmorRegistry.getSuitModel(item, contextModel.rightLeg, contextModel.leftLeg);
+        if (model == null) return;
+        ResourceLocation base = item.getTexture();
+        int packedColor = 0xFF000000 | item.getColor(stack);
+        model.renderToBuffer(matrices, vertexConsumers.getBuffer(model.renderType(base)), light, OverlayTexture.NO_OVERLAY, packedColor);
+        ResourceLocation overlay = item.getOverlayTexture();
+        if (overlay != null) {
+            model.renderToBuffer(matrices, vertexConsumers.getBuffer(model.renderType(overlay)), light, OverlayTexture.NO_OVERLAY, 0xFFFFFFFF);
+        }
     }
 }