+++ neoforge/src/main/java/net/satisfy/candlelight/forge/client/extensions/CandlelightChestplateExtensions.java
@@ -1,0 +1,21 @@
+package net.satisfy.candlelight.forge.client.extensions;
+
+import net.minecraft.client.model.HumanoidModel;
+import net.minecraft.client.model.Model;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.item.ItemStack;
+import net.neoforged.neoforge.client.extensions.common.IClientItemExtensions;
+import net.satisfy.candlelight.core.item.CandlelightChestItem;
+import net.satisfy.candlelight.core.registry.ArmorRegistry;
+import org.jetbrains.annotations.NotNull;
+
+public class CandlelightChestplateExtensions implements IClientItemExtensions {
+    @Override
+    public @NotNull Model getGenericArmorModel(@NotNull LivingEntity entity, @NotNull ItemStack stack, @NotNull EquipmentSlot slot, @NotNull HumanoidModel<?> original) {
+        if (slot == EquipmentSlot.CHEST && stack.getItem() instanceof CandlelightChestItem chest) {
+            return ArmorRegistry.getChestplateModel(chest, original.body, original.leftArm, original.rightArm, original.leftLeg, original.rightLeg);
+        }
+        return original;
+    }
+}