+++ neoforge/src/main/java/net/satisfy/candlelight/forge/client/extensions/GoldRingExtensions.java
@@ -1,0 +1,22 @@
+package net.satisfy.candlelight.forge.client.extensions;
+
+import net.minecraft.client.model.HumanoidModel;
+import net.minecraft.client.model.Model;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.item.ArmorMaterial;
+import net.minecraft.world.item.ItemStack;
+import net.neoforged.neoforge.client.extensions.common.IClientItemExtensions;
+import org.jetbrains.annotations.NotNull;
+
+public class GoldRingExtensions implements IClientItemExtensions {
+    @Override
+    public @NotNull Model getGenericArmorModel(LivingEntity entity, ItemStack stack, EquipmentSlot slot, HumanoidModel<?> original) {
+        return original;
+    }
+
+    @Override
+    public int getArmorLayerTintColor(ItemStack stack, LivingEntity entity, ArmorMaterial.Layer layer, int layerIdx, int fallbackColor) {
+        return 0;
+    }
+}